+++ src/main/java/org/leadpony/joy/internal/JsonGeneratorFactoryImpl.java
@@ -37,20 +37,28 @@ class JsonGeneratorFactoryImpl extends ConfigurableFactory implements JsonGenera
 
     private static final String[] SUPPORTED_PROPERTIES = {
         JsonGenerator.PRETTY_PRINTING,
-        JsonGenerator.INDENTATION_SIZE
+        JsonGenerator.INDENTATION_SIZE,
+        JsonGenerator.TAB_INDENTATION
     };
 
-    private static final int DEFAULT_INDENT_SIZE = 4;
+    private static final int DEFAULT_INDENTATION_SIZE = 4;
 
     private final boolean prettyPrinting;
-    private final int indentSize;
+    private final char indentationChar;
+    private final int indentationSize;
     private final CharBufferFactory bufferFactory;
 
     JsonGeneratorFactoryImpl(Map<String, ?> config, CharBufferFactory bufferFactory) {
         super(config, SUPPORTED_PROPERTIES);
-        this.prettyPrinting = containsProperty(JsonGenerator.PRETTY_PRINTING);
-        this.indentSize = getPropertyValue(JsonGenerator.INDENTATION_SIZE, DEFAULT_INDENT_SIZE);
         this.bufferFactory = bufferFactory;
+        this.prettyPrinting = containsProperty(JsonGenerator.PRETTY_PRINTING);
+        if (containsProperty(JsonGenerator.TAB_INDENTATION)) {
+            this.indentationChar = '\t';
+            this.indentationSize = 1;
+        } else {
+            this.indentationChar = ' ';
+            this.indentationSize = getPropertyValue(JsonGenerator.INDENTATION_SIZE, DEFAULT_INDENTATION_SIZE);
+        }
     }
 
     @Override
@@ -76,7 +84,7 @@ class JsonGeneratorFactoryImpl extends ConfigurableFactory implements JsonGenera
 
     private JsonGenerator createConfiguredGenerator(Writer writer) {
         if (prettyPrinting) {
-            return new PrettyJsonGenerator(writer, bufferFactory, indentSize);
+            return new PrettyJsonGenerator(writer, bufferFactory, indentationChar, indentationSize);
         } else {
             return new CompactJsonGenerator(writer, bufferFactory);
         }