+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- New configuration properties `INDENTATION_SIZE` and `TAB_INDENTATION` to the extended interface `JsonGenerator`.
+
 ## 1.0.1 - 2019-07-28
 ### Fixed
 - `getConfigInUse()` methods now return only properties recognized by the factory.